refactor: make `agent` a package
